fix(dequantize_rowwise): honest torch baseline — no int8->fp32 materialisation
x.to(torch.float32) copied the int8 input into a 4x-sized fp32 tensor
before the multiply; dropping it lets the single fused elementwise
kernel read int8 directly (type promotion happens in-kernel). The
trailing fp16 cast stays — it is the operator's output contract.

Full default + autotune rerun (20/20 verify). Sweep-max: torch 0.254
-> 0.186 ms; speedup_cutile 10.9 -> 8.0. DSL numbers unchanged.

+++ benchmarks/operators/dequantize_rowwise/impl_torch.py
@@ -3,4 +3,7 @@ import torch
 
 
 def run(x: torch.Tensor, state_x: torch.Tensor, **kwargs):
-    return (state_x.unsqueeze(1) * x.to(torch.float32) * (1.0 / 127.0)).to(torch.float16)
+    # int8 x is read directly by the fused elementwise kernel (type promotion
+    # to fp32 happens in-kernel) — the old x.to(torch.float32) materialised a
+    # 4x-sized intermediate first. The final fp16 cast is the output contract.
+    return (state_x.unsqueeze(1) * x * (1.0 / 127.0)).to(torch.float16)
